make address validation field public, move test to other package to be sure client code can be invoked from other package
